CORE-767: Move estimateLimit
